+++ config/config.go
@@ -26,6 +26,7 @@ type Config struct {
 	Include            []string
 	Exclude            []string
 	Order              []string
+	PrefixOrder        bool
 	Concurrency        int
 	Verbose            bool
 	FollowSymlinks     bool