+++ internal/align/strategy.go
@@ -13,6 +13,8 @@ type Options struct {
 	Schema *Schema
 
 	Types map[string]struct{}
+
+	PrefixOrder bool
 }
 
 type Schema struct {