+++ tests/cases/module/prefix_order/in.tf
@@ -1,0 +1,7 @@
+module "example" {
+  providers = {
+    b = aws.b
+    a = aws.a
+  }
+  source = "./m"
+}