+++ tests/cases/module/prefix_order/out.tf
@@ -1,0 +1,7 @@
+module "example" {
+  source = "./m"
+  providers = {
+    a = aws.a
+    b = aws.b
+  }
+}